Refactor test folders for secp256k1 (#364)
