+++ frost-secp256k1/src/tests/signing_tests.rs
@@ -1,0 +1,38 @@
+use lazy_static::lazy_static;
+use rand::thread_rng;
+use serde_json::Value;
+
+use crate::*;
+
+lazy_static! {
+    pub static ref VECTORS: Value =
+        serde_json::from_str(include_str!("../../tests/test_helpers/vectors.json").trim())
+            .expect("Test vector is valid JSON");
+    pub static ref VECTORS_BIG_IDENTIFIER: Value = serde_json::from_str(
+        include_str!("../../tests/test_helpers/vectors-big-identifier.json").trim()
+    )
+    .expect("Test vector is valid JSON");
+}
+
+/// This is testing that Shamir's secret sharing to compute and arbitrary
+/// value is working.
+#[test]
+fn check_share_generation_secp256k1_sha256() {
+    let rng = thread_rng();
+    frost_core::tests::check_share_generation::<Secp256K1Sha256, _>(rng);
+}
+
+#[test]
+fn check_sign_with_test_vectors() {
+    frost_core::tests::vectors::check_sign_with_test_vectors::<Secp256K1Sha256>(&VECTORS);
+    frost_core::tests::vectors::check_sign_with_test_vectors::<Secp256K1Sha256>(
+        &VECTORS_BIG_IDENTIFIER,
+    );
+}
+
+#[test]
+fn check_sign_with_dealer() {
+    let rng = thread_rng();
+
+    frost_core::tests::check_sign_with_dealer::<Secp256K1Sha256, _>(rng);
+}
